requirements.txt
pycryptodomex==3.19.0
python-constraint==1.4.0
pyzipper==0.3.6
